Moving forward with new ConsentBanner
