+++ src/ConsentBanner/index.ts
@@ -1,5 +1,1 @@
-import { ConsentBanner } from "./ConsentBanner";
-export * from "./ConsentBanner";
-export { addConsentBannerTranslations } from "./i18n";
-
-export default ConsentBanner;
+export type { RegisterConsentBannerFinality } from "./types";
