+++ src/tools/deepCopy.ts
@@ -1,0 +1,4 @@
+/** Assert obj is serializable */
+export function deepCopy<T>(obj: T): T {
+    return JSON.parse(JSON.stringify(obj));
+}
