+++ test/integration/cra/src/gdpr.ts
@@ -3,8 +3,9 @@ import { createConsentBanner } from "@codegouvfr/react-dsfr/ConsentBannerNext";
 
 
 export const { ConsentBanner, useFinalitiesConsent } = createConsentBanner<
-    "analytics" |
-    "statistics.traffic" | "statistics.deviceInfo" |
-    "personalization" | 
-    "advertising"
+    | "analytics" 
+    | "statistics.traffic" 
+    | "statistics.deviceInfo" 
+    | "personalization" 
+    | "advertising"
 >();
